Turn gulp task into a phase.
